Made List DTo genric

+++ Classroom/src/main/java/com/Classroom/Classroom/Service/AbsentService.java
@@ -12,5 +12,5 @@ public interface AbsentService {
     List<StudentAbsentDto> getAllAbsentDetails();
 
     List<StudentDto> getAbsentsOnDate(LocalDate specificDate);
-    String markDownAbsent(List<Integer> absenteesNumbers );
+    String markDownAbsent(List<Long> absenteesNumbers );
 }
